webui/requirements.txt
flask
werkzeug
pdf-craft
onnxruntime==1.21.0
torch>=2.0.0
python-docx>=0.8.11
markdown>=3.4.0
reportlab>=3.6.0
requests>=2.25.0
psutil>=5.9.0
PyMuPDF>=1.23.0

# 高性能OCR引擎
pytesseract>=0.3.10
Pillow>=9.0.0
opencv-python>=4.8.0
easyocr>=1.7.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
rapidocr-onnxruntime>=1.3.0

# 文档处理增强
pdfplumber>=0.9.0
pdf2image>=1.16.0
numpy>=1.21.0
scikit-image>=0.19.0